+++ impl_core/src/main/java/io/opencensus/implcore/tags/propagation/TagPropagationComponentImpl.java
@@ -18,6 +18,7 @@ package io.opencensus.implcore.tags.propagation;
 
 import io.opencensus.implcore.internal.CurrentState;
 import io.opencensus.tags.propagation.TagContextBinarySerializer;
+import io.opencensus.tags.propagation.TagContextTextFormat;
 import io.opencensus.tags.propagation.TagPropagationComponent;
 
 /** Implementation of {@link TagPropagationComponent}. */
@@ -31,5 +32,10 @@ public final class TagPropagationComponentImpl extends TagPropagationComponent {
   @Override
   public TagContextBinarySerializer getBinarySerializer() {
     return tagContextBinarySerializer;
+  }
+
+  @Override
+  public TagContextTextFormat getCorrelationContextFormat() {
+    throw new UnsupportedOperationException("not implemented");
   }
 }